feat. 엔티티,DTO 필드 타입 수정으로 null 처리 개선

+++ src/main/java/clofi/runningplanet/member/dto/response/ProfileResponse.java
@@ -9,17 +9,17 @@ public record ProfileResponse(
 
 	Gender gender,
 
-	int age,
+	Integer age,
 
 	String profileImg,
 
-	int runScore,
+	Integer runScore,
 
 	AvgPace avgPace,
 
-	double avgDistance,
+	Integer avgDistance,
 
-	double totalDistance,
+	int totalDistance,
 
 	String myCrew
 ) {
